Move unit tests and unify CMakeLists.txt
